update game history design implementation + flat features
