--- src/features/game/components/gamescreen/components/GameScreen.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { GameContainer } from '../../../../../ui/containers/GameContainer';
-
-export const GameScreen: React.FC = () => {
-  return (
-    <GameContainer>
-      <div>Jeu</div>
-    </GameContainer>
-  );
-};
