+++ src/features/game/rules/RuleLine.tsx
