+++ src/features/game/scores/PlayerScores.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import humanPlayer from '../../../../../assets/images/avatar-human.svg';
-import npcPlayer from '../../../../../assets/images/avatar-robot.svg';
-import { PlayerSection } from './playersection/PlayerSection';
+import humanPlayer from '../../../assets/images/avatar-human.svg';
+import npcPlayer from '../../../assets/images/avatar-robot.svg';
+import { PlayerSection } from './PlayerSection';
 
 export const PlayerScores: React.FC = () => {
   return (
@@ -12,14 +12,14 @@ export const PlayerScores: React.FC = () => {
         avatar={humanPlayer}
         altAvatar="avatar human player"
         scores={0}
-        isComputer={false}
+        isOpponent={false}
       />
       <PlayerSection
         playerName="J-Ordi"
         avatar={npcPlayer}
         altAvatar="avatar npc player"
         scores={0}
-        isComputer
+        isOpponent
       />
     </div>
   );
