+++ src/features/game/screen/GameScreen.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { GameContainer } from '../../../ui/containers/GameContainer';
+
+export const GameScreen: React.FC = () => {
+  return <GameContainer></GameContainer>;
+};
